requirements.txt
requests>=2.25.0
python-dotenv>=0.15.0
loguru>=0.7.0
questionary>=1.10.0
python-gitlab>=3.15.0
pyyaml>=6.0.0
pydantic>=2.0.0
typer>=0.9.0
